Fixed POM; All Runtime DataStoreProviders now detected
The project POM was autogenerated as part of Netbeans maven javafx project. The exec plugin execution section had outdated javafxpackager commands and the run command did not pass in the classpath so the runtime libs never got included in the classpath. As a result SIS was unable to pick up storage providers

The current implementation uses default exec plugin routines which work well for javafx projects.