Removed unused dependencies|Fixed empty south and west values
GMapxFX, ReactFX and EasyBind have been removed as they are no longer used

South and North bounds in GeographicExtentBox were rendered empty for 90°S and 180°W.
The spinner value factory was set with just min and max, which causes spinner to default to min value.
Then angle format was set on these fields. So when an extent covering the whole world was set, south and north values didnt fire a "change" event so angle format would not format them.